+++ scrapy_zyte_api/handler.py
@@ -23,6 +23,27 @@ from .utils import USER_AGENT
 logger = logging.getLogger(__name__)
 
 
+def _body_max_size_exceeded(
+    body_size: int,
+    warnsize: Optional[int],
+    maxsize: Optional[int],
+    request_url: str,
+) -> bool:
+    if warnsize and body_size > warnsize:
+        logger.warning(
+            f"Actual response size {body_size} larger than "
+            f"download warn size {warnsize} in request {request_url}."
+        )
+
+    if maxsize and body_size > maxsize:
+        logger.warning(
+            f"Dropping the response for {request_url}: actual response size "
+            f"{body_size} larger than download max size {maxsize}."
+        )
+        return True
+    return False
+
+
 def _truncate_str(obj, index, text, limit):
     if len(text) <= limit:
         return
@@ -67,8 +88,8 @@ class _ScrapyZyteAPIBaseDownloadHandler:
             # We keep the client in the crawler object to prevent multiple,
             # duplicate clients with the same settings to be used.
             # https://github.com/scrapy-plugins/scrapy-zyte-api/issues/58
-            crawler.zyte_api_client = client
-        self._client: AsyncZyteAPI = crawler.zyte_api_client
+            crawler.zyte_api_client = client  # type: ignore[attr-defined]
+        self._client: AsyncZyteAPI = crawler.zyte_api_client  # type: ignore[attr-defined]
         logger.info("Using a Zyte API key starting with %r", self._client.api_key[:7])
         verify_installed_reactor(
             "twisted.internet.asyncioreactor.AsyncioSelectorReactor"
@@ -83,6 +104,7 @@ class _ScrapyZyteAPIBaseDownloadHandler:
         )
         self._param_parser = _ParamParser(crawler)
         self._retry_policy = _load_retry_policy(settings)
+        assert crawler.stats
         self._stats = crawler.stats
         self._must_log_request = settings.getbool("ZYTE_API_LOG_REQUESTS", False)
         self._truncate_limit = settings.getint("ZYTE_API_LOG_REQUESTS_TRUNCATE", 64)
@@ -92,6 +114,9 @@ class _ScrapyZyteAPIBaseDownloadHandler:
                 f"({self._truncate_limit}) is invalid. It must be 0 or a "
                 f"positive integer."
             )
+        self._default_maxsize = settings.getint("DOWNLOAD_MAXSIZE")
+        self._default_warnsize = settings.getint("DOWNLOAD_WARNSIZE")
+
         crawler.signals.connect(self.engine_started, signal=signals.engine_started)
         self._crawler = crawler
         self._fallback_handler = None
@@ -105,6 +130,7 @@ class _ScrapyZyteAPIBaseDownloadHandler:
         self._session = self._client.session(trust_env=self._trust_env)
         if not self._cookies_enabled:
             return
+        assert self._crawler.engine
         for middleware in self._crawler.engine.downloader.middleware.middlewares:
             if isinstance(middleware, self._cookie_mw_cls):
                 self._cookie_jars = middleware.jars
@@ -231,7 +257,18 @@ class _ScrapyZyteAPIBaseDownloadHandler:
         finally:
             self._update_stats(api_params)
 
-        return _process_response(api_response, request, self._cookie_jars)
+        response = _process_response(
+            api_response=api_response, request=request, cookie_jars=self._cookie_jars
+        )
+        if response and _body_max_size_exceeded(
+            len(response.body),
+            self._default_warnsize,
+            self._default_maxsize,
+            request.url,
+        ):
+            return None
+
+        return response
 
     def _process_request_error(self, request, error):
         detail = (error.parsed.data or {}).get("detail", error.message)
@@ -240,6 +277,9 @@ class _ScrapyZyteAPIBaseDownloadHandler:
             f"type={error.parsed.type!r}, request_id={error.request_id!r}) "
             f"while processing URL ({request.url}): {detail}"
         )
+        assert self._crawler
+        assert self._crawler.engine
+        assert self._crawler.spider
         for status, error_type, close_reason in (
             (401, "/auth/key-not-found", "zyte_api_bad_key"),
             (403, "/auth/account-suspended", "zyte_api_suspended_account"),