+++ tests/test_providers.py
@@ -1,5 +1,5 @@
-import sys
 from collections import defaultdict
+from typing import Annotated
 
 import pytest
 
@@ -259,13 +259,8 @@ async def test_provider_params_remove_unused_options(mockserver):
     )
 
 
-@pytest.mark.skipif(
-    sys.version_info < (3, 9), reason="No Annotated support in Python < 3.9"
-)
 @ensureDeferred
 async def test_provider_extractfrom(mockserver):
-    from typing import Annotated
-
     @attrs.define
     class AnnotatedProductPage(BasePage):
         product: Annotated[Product, ExtractFrom.httpResponseBody]
@@ -295,13 +290,8 @@ async def test_provider_extractfrom(mockserver):
     )
 
 
-@pytest.mark.skipif(
-    sys.version_info < (3, 9), reason="No Annotated support in Python < 3.9"
-)
 @ensureDeferred
 async def test_provider_extractfrom_double(mockserver, caplog):
-    from typing import Annotated
-
     @attrs.define
     class AnnotatedProductPage(BasePage):
         product: Annotated[Product, ExtractFrom.httpResponseBody]
@@ -322,13 +312,8 @@ async def test_provider_extractfrom_double(mockserver, caplog):
     assert "Multiple different extractFrom specified for product" in caplog.text
 
 
-@pytest.mark.skipif(
-    sys.version_info < (3, 9), reason="No Annotated support in Python < 3.9"
-)
 @ensureDeferred
 async def test_provider_extractfrom_override(mockserver):
-    from typing import Annotated
-
     @attrs.define
     class AnnotatedProductPage(BasePage):
         product: Annotated[Product, ExtractFrom.httpResponseBody]
@@ -359,13 +344,8 @@ async def test_provider_extractfrom_override(mockserver):
     )
 
 
-@pytest.mark.skipif(
-    sys.version_info < (3, 9), reason="No Annotated support in Python < 3.9"
-)
 @ensureDeferred
 async def test_provider_geolocation(mockserver):
-    from typing import Annotated
-
     @attrs.define
     class GeoProductPage(BasePage):
         product: Product
@@ -385,9 +365,6 @@ async def test_provider_geolocation(mockserver):
     assert item["product"].name == "Product name (country DE)"
 
 
-@pytest.mark.skipif(
-    sys.version_info < (3, 9), reason="No Annotated support in Python < 3.9"
-)
 @ensureDeferred
 async def test_provider_geolocation_unannotated(mockserver, caplog):
     @attrs.define
@@ -414,9 +391,6 @@ custom_attrs_input = {
 }
 
 
-@pytest.mark.skipif(
-    sys.version_info < (3, 9), reason="No Annotated support in Python < 3.9"
-)
 @pytest.mark.parametrize(
     "annotation",
     [
@@ -428,8 +402,6 @@ custom_attrs_input = {
 )
 @ensureDeferred
 async def test_provider_custom_attrs(mockserver, annotation):
-    from typing import Annotated
-
     @attrs.define
     class CustomAttrsPage(BasePage):
         product: Product
@@ -468,13 +440,8 @@ async def test_provider_custom_attrs(mockserver, annotation):
     )
 
 
-@pytest.mark.skipif(
-    sys.version_info < (3, 9), reason="No Annotated support in Python < 3.9"
-)
 @ensureDeferred
 async def test_provider_custom_attrs_values(mockserver):
-    from typing import Annotated
-
     @attrs.define
     class CustomAttrsPage(BasePage):
         product: Product
@@ -540,6 +507,8 @@ async def test_provider_any_response_only(mockserver):
         response: AnyResponse
 
     class ZyteAPISpider(Spider):
+        url: str
+
         def start_requests(self):
             yield Request(self.url, callback=self.parse_)
 
@@ -567,6 +536,8 @@ async def test_provider_any_response_http_response_param(mockserver):
         response: AnyResponse
 
     class ZyteAPISpider(Spider):
+        url: str
+
         def start_requests(self):
             yield Request(self.url, callback=self.parse_)
 
@@ -595,6 +566,8 @@ async def test_provider_any_response_browser_html_param(mockserver):
         response: AnyResponse
 
     class ZyteAPISpider(Spider):
+        url: str
+
         def start_requests(self):
             yield Request(self.url, callback=self.parse_)
 
@@ -623,6 +596,8 @@ async def test_provider_any_response_product(mockserver):
         product: Product
 
     class ZyteAPISpider(Spider):
+        url: str
+
         def start_requests(self):
             yield Request(self.url, callback=self.parse_)
 
@@ -652,6 +627,8 @@ async def test_provider_any_response_product_extract_from_browser_html(mockserve
         product: Product
 
     class ZyteAPISpider(Spider):
+        url: str
+
         def start_requests(self):
             yield Request(self.url, callback=self.parse_)
 
@@ -684,6 +661,8 @@ async def test_provider_any_response_product_item_extract_from_browser_html(mock
         response: AnyResponse
 
     class ZyteAPISpider(Spider):
+        url: str
+
         def start_requests(self):
             yield Request(self.url, callback=self.parse_)
 
@@ -718,6 +697,8 @@ async def test_provider_any_response_product_extract_from_browser_html_2(mockser
         product: Product
 
     class ZyteAPISpider(Spider):
+        url: str
+
         def start_requests(self):
             yield Request(self.url, callback=self.parse_)
 
@@ -754,6 +735,8 @@ async def test_provider_any_response_product_extract_from_http_response(mockserv
         product: Product
 
     class ZyteAPISpider(Spider):
+        url: str
+
         def start_requests(self):
             yield Request(self.url, callback=self.parse_)
 
@@ -788,6 +771,8 @@ async def test_provider_any_response_product_options_empty(mockserver):
         product: Product
 
     class ZyteAPISpider(Spider):
+        url: str
+
         def start_requests(self):
             yield Request(self.url, callback=self.parse_)
 
@@ -824,6 +809,8 @@ async def test_provider_any_response_product_extract_from_http_response_2(mockse
         product: Product
 
     class ZyteAPISpider(Spider):
+        url: str
+
         def start_requests(self):
             yield Request(self.url, callback=self.parse_)
 
@@ -859,6 +846,8 @@ async def test_provider_any_response_browser_html(mockserver):
         html: BrowserHtml
 
     class ZyteAPISpider(Spider):
+        url: str
+
         def start_requests(self):
             yield Request(self.url, callback=self.parse_)
 
@@ -885,6 +874,8 @@ async def test_provider_any_response_browser_response(mockserver):
         browser_response: BrowserResponse
 
     class ZyteAPISpider(Spider):
+        url: str
+
         def start_requests(self):
             yield Request(self.url, callback=self.parse_)
 
@@ -912,6 +903,8 @@ async def test_provider_any_response_browser_html_response(mockserver):
         html: BrowserHtml
 
     class ZyteAPISpider(Spider):
+        url: str
+
         def start_requests(self):
             yield Request(self.url, callback=self.parse_)
 
@@ -939,6 +932,8 @@ async def test_provider_any_response_http_response(mockserver):
         http_response: HttpResponse
 
     class ZyteAPISpider(Spider):
+        url: str
+
         def start_requests(self):
             yield Request(self.url, callback=self.parse_)
 
@@ -970,6 +965,8 @@ async def test_provider_any_response_browser_http_response(mockserver):
         http_response: HttpResponse
 
     class ZyteAPISpider(Spider):
+        url: str
+
         def start_requests(self):
             yield Request(self.url, callback=self.parse_)
 
@@ -1008,6 +1005,8 @@ async def test_provider_any_response_http_response_multiple_pages(mockserver):
         response: AnyResponse
 
     class ZyteAPISpider(Spider):
+        url: str
+
         def start_requests(self):
             yield Request(self.url, callback=self.parse_)
 
@@ -1042,6 +1041,8 @@ async def test_provider_any_response_http_browser_response_multiple_pages(mockse
         response: AnyResponse
 
     class ZyteAPISpider(Spider):
+        url: str
+
         def start_requests(self):
             yield Request(self.url, callback=self.parse_)
 
@@ -1069,6 +1070,8 @@ async def test_provider_any_response_http_browser_response_multiple_pages(mockse
 @ensureDeferred
 async def test_screenshot(mockserver):
     class ZyteAPISpider(Spider):
+        url: str
+
         def start_requests(self):
             yield Request(self.url, callback=self.parse_)
 
@@ -1086,13 +1089,8 @@ async def test_screenshot(mockserver):
     assert item["screenshot"].body == b"screenshot-body-contents"
 
 
-@pytest.mark.skipif(
-    sys.version_info < (3, 9), reason="No Annotated support in Python < 3.9"
-)
 @ensureDeferred
 async def test_provider_actions(mockserver, caplog):
-    from typing import Annotated
-
     @attrs.define
     class ActionProductPage(BasePage):
         product: Product